requirements.txt
django 
djangorestframework 
djangorestframework-simplejwt 
django-channels
channels-redis
daphne
psycopg2
python-dotenv
stripe